# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Local personal memory server: episodic (OpenSearch) + knowledge (Neo4j) derived from canonical hot JSON, archived to S3.",
        "title": "memory-mcp v2 API",
        "contact": {},
        "version": "2.0"
    },
    "host": "127.0.0.1:8420",
    "basePath": "/",
    "paths": {
...
        "/v1/{ws}/{team}/{proj}/episodes": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "episodic"
                ],
                "summary": "Append an episodic record (hot write + best-effort index)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace",
                        "name": "ws",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "team",
                        "name": "team",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project",
                        "name": "proj",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "record",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpserver.CreateEpisodeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/episode.AppendResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    }
                }
            }
        },
        "/v1/{ws}/{team}/{proj}/episodes/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "episodic"
                ],
                "summary": "Full-text (nori) search over hot episodes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace",
                        "name": "ws",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "team",
                        "name": "team",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project",
                        "name": "proj",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "query text",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 lower bound",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 upper bound",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "comma-separated kinds",
                        "name": "kinds",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/episode.Hit"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    },
                    "503": {
                        "description": "index unavailable (§5)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    }
                }
            }
        },
        "/v1/{ws}/{team}/{proj}/episodes/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "episodic"
                ],
                "summary": "Fetch one episode by id (hot first, cold archive fallback)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace",
                        "name": "ws",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "team",
                        "name": "team",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project",
                        "name": "proj",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "episode ULID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/episode.Record"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apierr.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "episode.Actor": {
            "type": "string",
            "enum": [
                "agent",
                "user",
                "system"
            ],
            "x-enum-varnames": [
                "ActorAgent",
                "ActorUser",
                "ActorSystem"
            ]
        },
        "episode.AppendResult": {
            "type": "object",
            "properties": {
                "degraded": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "record": {
                    "$ref": "#/definitions/episode.Record"
                }
            }
        },
        "episode.Hit": {
            "type": "object",
            "properties": {
                "excerpt": {
                    "description": "Excerpt is a highlighted fragment of Text, not the full body.",
                    "type": "string"
                },
                "record": {
                    "$ref": "#/definitions/episode.Record"
                },
                "score": {
                    "type": "number"
                }
            }
        },
        "episode.Kind": {
            "type": "string",
            "enum": [
                "event",
                "conversation",
                "decision",
                "observation",
                "document_chunk"
            ],
            "x-enum-varnames": [
                "KindEvent",
                "KindConversation",
                "KindDecision",
                "KindObservation",
                "KindDocumentChunk"
            ]
        },
        "episode.Record": {
            "type": "object",
            "properties": {
                "actor": {
                    "description": "Actor is agent|user|system.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/episode.Actor"
                        }
                    ]
                },
                "consolidated": {
                    "description": "Consolidated marks the record as distilled into knowledge; a\nprecondition for cold archival (§3). Never inferred by the server: it is\nset only where the agent states the distillation itself, by naming this\nrecord in the provenance of a POST .../knowledge/nodes (§3, §0\nprinciple 2).",
                    "type": "boolean"
                },
                "entities": {
                    "description": "Entities are normalized entity names mentioned by this record.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "description": "ID is a ULID; immutable forever (provenance anchor).",
                    "type": "string"
                },
                "kind": {
                    "description": "Kind is one of the values accepted by ValidKind.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/episode.Kind"
                        }
                    ]
                },
                "last_recalled": {
                    "type": "string"
                },
                "occurred_at": {
                    "description": "OccurredAt is the event time (drives TTL aging, §3.1).",
                    "type": "string"
                },
                "recall_count": {
                    "description": "RecallCount / LastRecalled track search hits. LastRecalled is RFC3339\nor \"\" when never recalled.",
                    "type": "integer"
                },
                "refs": {
                    "description": "Refs is set only for kind=document_chunk.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/episode.Refs"
                        }
                    ]
                },
                "text": {
                    "description": "Text is the body — the nori full-text indexing target.",
                    "type": "string"
                }
            }
        },
        "episode.Refs": {
            "type": "object",
            "properties": {
                "chunk_seq": {
                    "type": "integer"
                },
                "doc_sha": {
                    "type": "string"
                }
            }
        },
...
        "httpserver.CreateEpisodeRequest": {
            "type": "object",
            "properties": {
                "actor": {
                    "$ref": "#/definitions/episode.Actor"
                },
                "entities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "kind": {
                    "$ref": "#/definitions/episode.Kind"
                },
                "occurred_at": {
                    "type": "string"
                },
                "refs": {
                    "$ref": "#/definitions/episode.Refs"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "httpserver.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/apierr.Error"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}